c_importer_product: split product importer files
