+++ connector_importer_product/components/product_product/__init__.py
@@ -1,0 +1,3 @@
+from . import importer
+from . import record_handler
+from . import mapper
